chore : 컬럼 데이터타입 수정  - postId long -> int로 변경

+++ src/main/java/blogservices/blogdevbackend/domain/comment/service/CommentService.java
@@ -9,7 +9,7 @@ public interface CommentService {
 
     List<CommentResponseDto> save(CommentRequestDto request);
 
-    List<CommentResponseDto> getComment(long postId);
+    List<CommentResponseDto> getComment(int postId);
 
     Long update(long commentId, CommentRequestDto request);
